requirements.txt
pyserial==3.5
mouse==0.7.1